Build new program structure
MainMixer is the main machinery with scenes and channels.
It yields to Playback

Playback puts audio from MainMixer in a buffer from where it
is played back either through discord-bot or local speakers

PyQt Ui has the widget files generated with pyuic5, they are
qml and should not be altered by hand.
They are loaded to main gui which applies logic and function
to the layout.

Controller is the middleman that handles interactions between
all the program elements.